requirements.txt
langchain==0.3.7
langchain-community==0.3.3
langchain-core==1.0.4
langchain-text-splitters==0.3.2
langchain-google-genai==1.0.3
langchain-groq==0.2.1
langchain-huggingface==1.0.1
faiss-cpu==1.12.0
sentence-transformers==5.1.2
transformers==4.57.1
huggingface-hub==0.36.0
torch==2.9.0
groq==0.33.0
google-generativeai==0.5.4
python-dotenv==1.2.1
pypdf==6.2.0
beautifulsoup4==4.12.3
requests==2.32.5
tqdm==4.67.1
numpy==1.26.4
scikit-learn==1.7.2
scipy==1.16.3
tenacity==8.5.0
typing-extensions==4.15.0
